apps/ocr-worker/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
paddleocr==2.7.0
paddlepaddle==2.6.2
pillow==10.1.0
numpy==1.26.4
pdf2image==1.17.0
